Rename note_encryption_vanilla/note_encryption_zsa to orchard_domain_vanilla/orchard_domain_zsa
